Wait for client to apply edit before moving cursor
This fixes a bug that was noticed in Calva. The server was sending both
workspace/applyEdit and window/showDocument simultaneously and
asynchronously. Sometimes Calva received and/or processed the
showDocument request before the applyEdit request, meaning the cursor
didn't end up in the correct place. To fix this we now wait for the
client to acknowledge the applyEdit command before sending the
showDocument command, ensuring they are processed in the correct order.

This also teaches the integration tests how to mock a client's response
to a server request.